requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
Flask-Cors==4.0.0
geographiclib==2.0
geopy==2.4.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
prettytable==3.10.0
protobuf==4.21.12
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
